OzonParseApp/requirements.txt
amqp==5.1.1
anyio==3.7.1
asgiref==3.7.2
attrs==23.1.0
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==41.0.2
Django==3.2
django-adminlte3==0.1.6
djangorestframework==3.14.0
exceptiongroup==1.1.2
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
kombu==5.3.1
mysqlclient==2.2.0
outcome==1.2.0
packaging==23.1
prompt-toolkit==3.0.39
protobuf==4.21.12
pycparser==2.21
pyOpenSSL==23.2.0
PySocks==1.7.1
python-dateutil==2.8.2
python-telegram-bot==20.4
pytz==2023.3
redis==4.6.0
requests==2.31.0
selenium==4.11.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.22.2
trio-websocket==0.10.3
tzdata==2023.3
undetected-chromedriver==3.5.0
urllib3==1.26.16
urllib3-secure-extra==0.1.0
vine==5.0.0
wcwidth==0.2.6
websockets==11.0.3
wsproto==1.2.0
xvfbwrapper==0.2.9
